backend/requirements.txt
alembic==1.6.5
fastapi==0.67.0
FastAPI-SQLAlchemy==0.2.1
pydantic==1.8.2
psycopg2-binary==2.9.1
uvicorn==0.14.0
fastapi-auth0==0.2.1
PyJWT==2.1.0
Faker==8.11.0
pytest==6.2.4
pytest-asyncio==0.15.1
databases==0.4.3
httpx==0.18.2
asgi-lifespan==1.0.1
